requirements.txt
numpy
selenium
tensorflow
pandas
matplotlib
ipykernel
scikit-learn
pyyaml